profil data 1

+++ Template/Admin/Users/mainimage.ctp
@@ -1,10 +1,10 @@
 
 <h1><?php echo $title ?></h1>
-<?php echo $this->Html->link('<i class="fa fa-arrow-circle-left"></i> Wróć do użytkownika </a>', ['controller'=>'users', 'action'=>'manage', $userID ], array('escape'=>false,'class'=>'btn default btn-xs blue')); ?>
+<?php echo $this->Html->link('<i class="fa fa-arrow-circle-left"></i> Wróć do użytkownika', ['controller'=>'users', 'action'=>'manage', $userID ], array('escape'=>false,'class'=>'btn default btn-xs blue')); ?>
 <br />
 <hr />
 <br />
-		<div class="portlet box green">
+		<div class="portlet box red">
 			<div class="portlet-title">
 				<div class="caption"><i class="fa fa-camera"></i><?php echo $subtitle ?></div>
 			</div>
@@ -21,9 +21,9 @@
 							<label class="control-label col-md-3">Avatar:</label>
 							<div class="controls  col-md-9">
 								<?php echo $this->Form->input('id',['type'=>'hidden','value'=>$userID,'class'=>'form-control', 'label'=>false]);?>
-								<?php echo $this->Form->input('mainImage',['value'=>$mainimage, 'class'=>'form-control', 'label'=>false]);?>
+								<?php echo $this->Form->input('mainImage',['type'=>'hidden','value'=>$mainimage, 'class'=>'form-control', 'label'=>false]);?>
 								<?php
-								if(file_exists($staticurl."profiles/".$userID."/".$mainimage)){
+								if($mainimage != '' && file_exists($staticurl."profiles/".$userID."/".$mainimage)){
 									echo '<img class="img-responsive" src="'.$staticurl.'profiles/'.$userID.'/'.$mainimage.'" alt="">';
 								}
 								else{
@@ -39,7 +39,7 @@
 									<div class="row">
 										<div class="col-md-offset-3 col-md-9">
 											<button type="submit" class="btn green"><i class="fa fa-check"></i> Submit</button>
-											<button type="button" class="btn default">Cancel</button>
+											<?php echo $this->Html->link('Anuluj', ['controller'=>'users', 'action'=>'manage', $userID ], array('escape'=>false,'class'=>'btn default')); ?>
 										</div>
 									</div>
 								</div>
@@ -53,5 +53,5 @@
 <br />
 <hr />
 <br />
-<?php echo $this->Html->link('<i class="fa fa-arrow-circle-left"></i> Wróć do użytkownika </a>', ['controller'=>'users', 'action'=>'manage', $userID ], array('escape'=>false,'class'=>'btn default btn-xs blue')); ?>
+<?php echo $this->Html->link('<i class="fa fa-arrow-circle-left"></i> Wróć do użytkownika ', ['controller'=>'users', 'action'=>'manage', $userID ], array('escape'=>false,'class'=>'btn default btn-xs blue')); ?>
